+++ packages/core/src/pages/[slug]/p.tsx
@@ -47,6 +47,7 @@ import type { PreviewData } from 'src/server/content/types'
 type StoreConfig = typeof storeConfig & {
   experimental: {
     revalidate?: number
+    revalidate404?: number
     enableClientOffer?: boolean
   }
 }
@@ -482,6 +483,8 @@ export const getStaticProps: GetStaticProps<
 
     return {
       notFound: true,
+      revalidate:
+        (storeConfig as StoreConfig).experimental.revalidate404 ?? 60 * 5, // 5 minutes
     }
   }